+++ __tests__/01_basic_spec.js
@@ -46,9 +46,9 @@ describe('basic spec', () => {
         </ReduxProvider>
       </StrictMode>
     );
-    const { getByText, container } = render(<App />);
+    const { getAllByText, container } = render(<App />);
     expect(container).toMatchSnapshot();
-    fireEvent.click(getByText('+1'));
+    fireEvent.click(getAllByText('+1')[0]);
     expect(container).toMatchSnapshot();
   });
 });