+++ src/index.js
@@ -1,4 +1,4 @@
 export { ReduxProvider } from './provider';
 export { useReduxDispatch } from './useReduxDispatch';
-export { useReduxState, useReduxStateSimple } from './useReduxState';
+export { useReduxState, useReduxStateRich, useReduxStateSimple } from './useReduxState';
 export { useReduxSelectors } from './useReduxSelectors';