+++ src/useReduxState.js
@@ -9,7 +9,12 @@ import { proxyState, proxyEqual } from 'proxyequal';
 
 import { ReduxStoreContext } from './provider';
 
-import { useIsomorphicLayoutEffect, useForceUpdate } from './utils';
+import {
+  useIsomorphicLayoutEffect,
+  useForceUpdate,
+  createDeepProxy,
+  isDeepChanged,
+} from './utils';
 
 const useTrapped = (state) => {
   const cacheRef = useRef({
@@ -27,7 +32,7 @@ const useTrapped = (state) => {
   return trapped;
 };
 
-export const useReduxState = () => {
+export const useReduxStateRich = () => {
   const forceUpdate = useForceUpdate();
   // redux store&state
   const store = useContext(ReduxStoreContext);
@@ -62,6 +67,49 @@ export const useReduxState = () => {
     return unsubscribe;
   }, [store]); // eslint-disable-line react-hooks/exhaustive-deps
   return trapped.state;
+};
+
+export const useReduxState = (opts = {}) => {
+  const forceUpdate = useForceUpdate();
+  const store = useContext(ReduxStoreContext);
+  const state = store.getState();
+  const affected = new WeakMap();
+  const lastTracked = useRef(null);
+  useIsomorphicLayoutEffect(() => {
+    lastTracked.current = {
+      state,
+      affected,
+      cache: new WeakMap(),
+      /* eslint-disable no-nested-ternary, indent, @typescript-eslint/indent */
+      assumeChangedIfNotAffected:
+        opts.unstable_forceUpdateForStateChange ? true
+      : opts.unstable_ignoreIntermediateObjectUsage ? false
+      : /* default */ null,
+      /* eslint-enable no-nested-ternary, indent, @typescript-eslint/indent */
+    };
+  });
+  useEffect(() => {
+    const callback = () => {
+      const nextState = store.getState();
+      const changed = isDeepChanged(
+        lastTracked.current.state,
+        nextState,
+        lastTracked.current.affected,
+        lastTracked.current.cache,
+        lastTracked.current.assumeChangedIfNotAffected,
+      );
+      if (changed) {
+        lastTracked.current.state = nextState;
+        forceUpdate();
+      }
+    };
+    // run once in case the state is already changed
+    callback();
+    const unsubscribe = store.subscribe(callback);
+    return unsubscribe;
+  }, [store]); // eslint-disable-line react-hooks/exhaustive-deps
+  const proxyCache = useRef(new WeakMap()); // per-hook proxyCache
+  return createDeepProxy(state, affected, proxyCache.current);
 };
 
 export const useReduxStateSimple = () => {